Moving cluster.hpp away
